feat: 모바일에서 인풋 키보드에서 다음 버튼이 표시되도록 하고 기능이 작동하도록 구현
