+++ docs/REQUIREMENTS_2.md
@@ -24,10 +24,10 @@
 
 - [x] 디바이스에서 뒤로가기를 하였을 때 페이지 기록으로 뒤로가는 것이 아닌 사이트 내에서 정말 뒤로 갈 것인지 묻는 기능 구현
 - [x] 처음에 오토포커스가 되어있어서 모달이 열린 상태로 숫자 키보드가 나오는 것을 방지하고자 카드사별을 선택한 후 첫번째 카드 번호 인풋으로 포커스하도록 수정하기
-- [ ] 디바이스 키보드에서 다음 버튼을 눌렀을 때 다음으로 이동하기
+- [x] 디바이스 키보드에서 다음 버튼을 눌렀을 때 다음으로 이동하기
 - [x] 카드 별칭 등록 화면 가운데 정렬하기
-- [x] 사용자 이름은 오로지 영어로만 입력 가능하도록 하게 하기 && 모바일에서 키보드 영어 대문자 자동 설정
-- [ ] 카드 별칭은 마음대로(이모티콘, 영어, 한글, 한자, 특수문자 가능) 입력 가능하나 20 글자 이하로만 입력 가능하도록 하기
+- [x] 사용자 이름은 오로지 영어로만 입력 가능하도록 하게 하기 && 모바일에서 키보드 영어 대문자를 자동 설정하도록 구현
+- [x] 카드 별칭은 마음대로(이모티콘, 영어, 한글, 한자, 특수문자 가능) 입력 가능하나 20 글자 이하로만 입력 가능하도록 하기
 - [x] 모달창이 열렸을 때 브라우저 탭이 하단에 있을 경우 잘려서 표시되는 것을 브라우저 탭이 열렸을 땐 브라우저 탭 위로 이동하도록 수정하기
 
 <!--
