+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { TAB_INDEX_INFO } from './constant';
+
 export const isNumber = (value: string) => {
   return !Number.isNaN(Number(value));
 };
@@ -47,4 +49,30 @@ export const lengthValidate = (maxLength: number) => (value: string) => {
 
 export const convertUpperCase = (text: string) => {
   return text.toUpperCase();
+};
+
+export const setNextInputFocus = (form: HTMLFormElement | null) => {
+  if (!form) return;
+
+  const active = document.activeElement as HTMLInputElement;
+  const currentInputIndex = Array.prototype.indexOf.call(form, active);
+
+  const totalLength = form.length;
+
+  for (let index = currentInputIndex + 1; index < totalLength; index += 1) {
+    const currentInput = form[index] as HTMLInputElement;
+
+    const { maxLength, value, tabIndex } = currentInput;
+
+    if (tabIndex === TAB_INDEX_INFO.dismiss) {
+      continue;
+    }
+    if (maxLength === value.length) {
+      continue;
+    }
+    setTimeout(() => {
+      currentInput.focus();
+    });
+    break;
+  }
 };
